+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 [full changelog](http://github.com/yolk/sum_sum/compare/v0.0.2...master)
 
+* Added SumSum#dump & SumSum.load methods to dump and load data to serializable objects (array, hash)
+
 ### 0.0.2 / 2011-01-27
 
 [full changelog](http://github.com/yolk/sum_sum/compare/v0.0.1...v0.0.2)